redmine: improve content of notification with journal_html
journal_html contains a html formatted details of updates done on a
redmine ticket. We want to use this if possible, but replace the html
markups with slack ones.

For the moment, only strong/i is handled, as I don't know the extent
of the HTML used by redmine. We'll see for other markups when we come
across them.